deploy separate files
